Move 'resolve kitfile' functions out of cmd/info and into pkg/lib
